Conflicts:
	wp-content/themes/ednc-roots/assets/app/sass/layouts/_posts.scss
	wp-content/themes/ednc-roots/assets/manifest.json
	wp-content/themes/ednc-roots/assets/public/css/main.css
	wp-content/themes/ednc-roots/assets/public/css/main.css.map
	wp-content/themes/ednc-roots/assets/public/css/main.min.css
	wp-content/themes/ednc-roots/assets/public/css/main.min.css.map
	wp-content/themes/ednc-roots/assets/public/imgs/burroughs-wellcome-fund.png
	wp-content/themes/ednc-roots/assets/public/imgs/jw-pope-foundation.png
	wp-content/themes/ednc-roots/assets/public/imgs/winston-salem-foundation.png

+++ wp-content/plugins/manual-image-crop/lib/ManualImageCrop.php
@@ -12,12 +12,12 @@ class ManualImageCrop {
 	 * Returns the instance of the class [Singleton]
 	 * @return ManualImageCrop
 	 */
-    public static function getInstance() {
-        if (self::$instance === null) {
-            self::$instance = new ManualImageCrop();
-        }
-        return self::$instance;
-    }
+	public static function getInstance() {
+		if (self::$instance === null) {
+			self::$instance = new ManualImageCrop();
+		}
+		return self::$instance;
+	}
 
 	private function __construct() {
 
@@ -64,16 +64,16 @@ class ManualImageCrop {
 	 * Adds link below "Remove featured image" in post editing form
 	 */
 	public function addCropFeatureImageEditorLink($content, $post) {
-		$content .= '<a id="micCropFeatureImage" class="thickbox mic-link" rel="crop" title="Manual Image Crop" href="' . admin_url( 'admin-ajax.php' ) . '?action=mic_editor_window&postId=' . get_post_thumbnail_id($post) . '">' . __('Crop featured image','microp') . '</a>
-<script>
-setInterval(function() {
-	if (jQuery(\'#remove-post-thumbnail\').is(\':visible\')) {
+		$content .= '<a id="micCropFeatureImage" class="thickbox mic-link" rel="crop" title="' . __('Manual Image Crop','microp') . '" href="' . admin_url( 'admin-ajax.php' ) . '?action=mic_editor_window&postId=' . get_post_thumbnail_id($post) . '">' . __('Crop featured image','microp') . '</a>
+		<script>
+		setInterval(function() {
+		if (jQuery(\'#remove-post-thumbnail\').is(\':visible\')) {
 		jQuery(\'#micCropFeatureImage\').show();
 	}else {
-		jQuery(\'#micCropFeatureImage\').hide();
+	jQuery(\'#micCropFeatureImage\').hide();
 	}
-}, 200);
-</script>';
+	}, 200);
+	</script>';
 		return $content;
 	}
 
@@ -81,17 +81,17 @@ setInterval(function() {
 	 * Adds link in the ligthbox media library
 	 */
 	public function addAttachementEditLink() { ?>
-	<script>
+<script>
 		var micEditAttachemtnLinkAdded = false;
 		var micEditAttachemtnLinkAddedInterval = 0;
-		jQuery(document).ready(function() {
+		jQuery(document).ready(function() {			
 			micEditAttachemtnLinkAddedInterval = setInterval(function() {
 				if (jQuery('.details .edit-attachment').length) {
 					try {
-						var mRegexp = /\?post=([0-9]+)/;
+						var mRegexp = /\?post=([0-9]+)/; 
 						var match = mRegexp.exec(jQuery('.details .edit-attachment').attr('href'));
 						jQuery('.crop-image-ml.crop-image').remove();
-						jQuery('.details .edit-attachment').after( '<a class="thickbox mic-link crop-image-ml crop-image" rel="crop" title="Manual Image Crop" href="' + ajaxurl + '?action=mic_editor_window&postId=' + match[1] + '"><?php _e('Crop Image','microp') ?></a>' );
+						jQuery('.details .edit-attachment').after( '<a class="thickbox mic-link crop-image-ml crop-image" rel="crop" title="<?php _e("Manual Image Crop","microp"); ?>" href="' + ajaxurl + '?action=mic_editor_window&postId=' + match[1] + '"><?php _e('Crop Image','microp') ?></a>' );
 					} catch (e) {
 						console.log(e);
 					}
@@ -101,7 +101,7 @@ setInterval(function() {
 					try {
 						var postId = jQuery('.attachment-details').attr('data-id');
 						jQuery('.button.crop-image-ml.crop-image').remove();
-						jQuery('.button.edit-attachment').after( ' <a class="thickbox mic-link crop-image-ml crop-image button" rel="crop" title="Manual Image Crop" href="' + ajaxurl + '?action=mic_editor_window&postId=' + postId + '"><?php _e('Crop Image','microp') ?></a>' );
+						jQuery('.button.edit-attachment').after( ' <a class="thickbox mic-link crop-image-ml crop-image button" rel="crop" title="<?php _e("Manual Image Crop","microp"); ?>" href="' + ajaxurl + '?action=mic_editor_window&postId=' + postId + '"><?php _e('Crop Image','microp') ?></a>' );
 					} catch (e) {
 						console.log(e);
 					}
@@ -109,15 +109,15 @@ setInterval(function() {
 			}, 500);
 		});
 	</script>
-	<?php
+<?php
 	}
 
 	/**
 	 * Adds link in the ligthbox media library
 	 */
 	public function addAfterUploadAttachementEditLink() {
-	?>
-	<script>
+		?>
+<script>
 		var micEditAttachemtnLinkAdded = false;
 		var micEditAttachemtnLinkAddedInterval = 0;
 		jQuery(document).ready(function() {
@@ -125,10 +125,10 @@ setInterval(function() {
 				if (jQuery('#media-items .edit-attachment').length) {
 					jQuery('#media-items .edit-attachment').each(function(i, k) {
 						try {
-							var mRegexp = /\?post=([0-9]+)/;
+							var mRegexp = /\?post=([0-9]+)/; 
 							var match = mRegexp.exec(jQuery(this).attr('href'));
 							if (!jQuery(this).parent().find('.edit-attachment.crop-image').length && jQuery(this).parent().find('.pinkynail').attr('src').match(/upload/g)) {
-								jQuery(this).after( '<a class="thickbox mic-link edit-attachment crop-image" rel="crop" title="Manual Image Crop" href="' + ajaxurl + '?action=mic_editor_window&postId=' + match[1] + '"><?php _e('Crop Image','microp') ?></a>' );
+								jQuery(this).after( '<a class="thickbox mic-link edit-attachment crop-image" rel="crop" title="<?php _e("Manual Image Crop","microp"); ?>" href="' + ajaxurl + '?action=mic_editor_window&postId=' + match[1] + '"><?php _e('Crop Image','microp') ?></a>' );
 							}
 						} catch (e) {
 							console.log(e);
@@ -138,7 +138,7 @@ setInterval(function() {
 			}, 500);
 		});
 	</script>
-  <?php
+<?php
 	}
 
 	/**
@@ -147,6 +147,12 @@ setInterval(function() {
 	public function cropImage() {
 		global $_wp_additional_image_sizes;
 
+		$dst_file_url = wp_get_attachment_image_src($_POST['attachmentId'], $_POST['editedSize']);
+
+		if (!$dst_file_url) {
+			exit;
+		}
+
 		$uploadsDir = wp_upload_dir();
 
 		// checks for ssl. wp_upload_dir does not handle ssl (ssl admin trips on this and subsequent ajax success to browser)
@@ -154,41 +160,38 @@ setInterval(function() {
 			$uploadsDir['baseurl'] = preg_replace('#^http://#i', 'https://', $uploadsDir['baseurl']);
 		}
 
-		$src_file_url = wp_get_attachment_image_src($_POST['attachmentId'], 'full');
 
-		if (!$src_file_url) {
-			echo json_encode (array('status' => 'error', 'message' => 'wrong attachment' ) );
-			exit;
+		if ( function_exists( '_load_image_to_edit_path' ) ) {
+			// this function is consider as private, but it return proper image path. Notice it is in function_exists condition
+			$src_file = _load_image_to_edit_path( $_POST['attachmentId'], 'full' );
+			$dst_file = _load_image_to_edit_path( $_POST['attachmentId'], $_POST['editedSize'] );
+		} else {
+			$src_file_url = wp_get_attachment_image_src( $_POST['attachmentId'], 'full' );
+
+			if ( ! $src_file_url ) {
+				echo json_encode( array( 'status' => 'error', 'message' => 'wrong attachment' ) );
+				exit;
+			}
+
+			$src_file = str_replace( $uploadsDir['baseurl'], $uploadsDir['basedir'], $src_file_url[0] );
+			$dst_file = str_replace( $uploadsDir['baseurl'], $uploadsDir['basedir'], $dst_file_url[0] );
 		}
-
-		if (is_ssl()) {
-			$src_file_url = preg_replace('#^http://#i', 'https://', $src_file_url);
-		}
-
-		$src_file = str_replace($uploadsDir['baseurl'], $uploadsDir['basedir'], $src_file_url[0]);
-
-		$dst_file_url = wp_get_attachment_image_src($_POST['attachmentId'], $_POST['editedSize']);
-
-		if (!$dst_file_url) {
-			exit;
-		}
-
-		if (is_ssl()) {
-			$dst_file_url = preg_replace('#^http://#i', 'https://', $dst_file_url);
-		}
-		
-		$dst_file = str_replace($uploadsDir['baseurl'], $uploadsDir['basedir'], $dst_file_url[0]);
-
 
 		//checks if the destination image file is present (if it's not, we want to create a new file, as the WordPress returns the original image instead of specific one)
 		if ($dst_file == $src_file) {
 			$attachmentData = wp_generate_attachment_metadata( $_POST['attachmentId'], $dst_file );
-
-			//new destination file path - replaces original file name with the correct one
-			$dst_file = str_replace( basename($attachmentData['file']), $attachmentData['sizes'][ $_POST['editedSize'] ]['file'], $dst_file);
-
+				
+			//overwrite with previous values
+			$prevAttachmentData = wp_get_attachment_metadata($_POST['attachmentId']);
+			if (isset($prevAttachmentData['micSelectedArea'])) {
+				$attachmentData['micSelectedArea'] = $prevAttachmentData['micSelectedArea'];
+			}
+				
 			//saves new path to the image size in the database
 			wp_update_attachment_metadata( $_POST['attachmentId'],  $attachmentData );
+				
+			//new destination file path - replaces original file name with the correct one
+			$dst_file = str_replace( basename($attachmentData['file']), $attachmentData['sizes'][ $_POST['editedSize'] ]['file'], $dst_file);
 
 			//retrieves the new url to file (needet to refresh the preview)
 			$dst_file_url = wp_get_attachment_image_src($_POST['attachmentId'], $_POST['editedSize']);
@@ -196,7 +199,7 @@ setInterval(function() {
 
 		//sets the destination image dimensions
 		if (isset($_wp_additional_image_sizes[$_POST['editedSize']])) {
-			$dst_w = min(intval($_wp_additional_image_sizes[$_POST['editedSize']]['width']), $_POST['select']['w'] * $_POST['previewScale']);;
+			$dst_w = min(intval($_wp_additional_image_sizes[$_POST['editedSize']]['width']), $_POST['select']['w'] * $_POST['previewScale']);
 			$dst_h = min(intval($_wp_additional_image_sizes[$_POST['editedSize']]['height']), $_POST['select']['h'] * $_POST['previewScale']);
 		} else {
 			$dst_w = min(get_option($_POST['editedSize'].'_size_w'), $_POST['select']['w'] * $_POST['previewScale']);
@@ -216,68 +219,92 @@ setInterval(function() {
 		$src_w = max(0, $_POST['select']['w']) * $_POST['previewScale'];
 		$src_h = max(0, $_POST['select']['h']) * $_POST['previewScale'];
 
+		$size = wp_get_image_editor( $src_file )->get_size();
+
+		$is_higher = ( $dst_h > $size["height"] );
+		$is_wider = ( $dst_w > $size["width"] );
+
+		if ( $is_higher || $is_wider ) {
+			if ( $is_higher ) {
+				$scale = $src_h / $size["height"];
+			} else {
+				$scale = $src_w / $size["width"];
+			}
+
+			$src_w = $src_w / $scale;
+			$src_h = $src_h / $scale;
+			$src_x = $src_x / $scale;
+			$src_y = $src_y / $scale;
+		}
+
 		//saves the selected area
 		$imageMetadata = wp_get_attachment_metadata($_POST['attachmentId']);
 		$imageMetadata['micSelectedArea'][$_POST['editedSize']] = array(
-																	'x' => $_POST['select']['x'],
-																	'y' => $_POST['select']['y'],
-																	'w' => $_POST['select']['w'],
-																	'h' => $_POST['select']['h'],
-																);
+				'x' => $_POST['select']['x'],
+				'y' => $_POST['select']['y'],
+				'w' => $_POST['select']['w'],
+				'h' => $_POST['select']['h'],
+				'scale' => $_POST['previewScale'],
+		);
 		wp_update_attachment_metadata($_POST['attachmentId'], $imageMetadata);
 
-		$quality = intval($_POST['mic_quality']);
+		$quality = isset($_POST['mic_quality']) ? intval($_POST['mic_quality']) : 60;
 
-        if ( function_exists('wp_get_image_editor') ) {
-            $img = wp_get_image_editor( $src_file );
+		if ( function_exists('wp_get_image_editor') ) {
+			$img = wp_get_image_editor( $src_file );
 
-            if ( ! is_wp_error( $img ) ) {
+			if ( ! is_wp_error( $img ) ) {
 
-            	$img->crop( $src_x, $src_y, $src_w, $src_h, $dst_w, $dst_h, false );
-                $img->set_quality( $quality );
-                $img->save($dst_file);
-            }else {
-            	echo json_encode (array('status' => 'error', 'message' => 'WP_ERROR: ' . $img->get_error_message() ) );
+				$img->crop( $src_x, $src_y, $src_w, $src_h, $dst_w, $dst_h, false );
+				$img->set_quality( $quality );
+				$saveStatus = $img->save( $dst_file );
+
+				if ( is_wp_error( $saveStatus ) ) {
+					echo json_encode( array( 'status' => 'error', 'message' => 'WP_ERROR: ' . $saveStatus->get_error_message() ) );
+					exit;
+				}
+			}else {
+				echo json_encode (array('status' => 'error', 'message' => 'WP_ERROR: ' . $img->get_error_message() ) );
 				exit;
-            }
-        } else {
-            //determines what's the image format
-            $ext = pathinfo($src_file, PATHINFO_EXTENSION);
-            if ($ext == "gif"){
-                $src_img = imagecreatefromgif($src_file);
-            } else if($ext =="png"){
-                $src_img = imagecreatefrompng($src_file);
-            } else {
-                $src_img = imagecreatefromjpeg($src_file);
-            }
+			}
+		} else {
+			//determines what's the image format
+			$ext = pathinfo($src_file, PATHINFO_EXTENSION);
+			if ($ext == "gif"){
+				$src_img = imagecreatefromgif($src_file);
+			} else if($ext =="png"){
+				$src_img = imagecreatefrompng($src_file);
+			} else {
+				$src_img = imagecreatefromjpeg($src_file);
+			}
 
-            if ($src_img === false ) {
-            	echo json_encode (array('status' => 'error', 'message' => 'PHP ERROR: Cannot create image from the source file' ) );
-            	exit;
-            }
+			if ($src_img === false ) {
+				echo json_encode (array('status' => 'error', 'message' => 'PHP ERROR: Cannot create image from the source file' ) );
+				exit;
+			}
 
-            $dst_img = imagecreatetruecolor($dst_w, $dst_h);
-            $resampleReturn  = imagecopyresampled($dst_img, $src_img, $dst_x, $dst_y, $src_x, $src_y, $dst_w, $dst_h, $src_w, $src_h);
+			$dst_img = imagecreatetruecolor($dst_w, $dst_h);
+			$resampleReturn  = imagecopyresampled($dst_img, $src_img, $dst_x, $dst_y, $src_x, $src_y, $dst_w, $dst_h, $src_w, $src_h);
 
-            if ($resampleReturn === false ) {
-            	echo json_encode (array('status' => 'error', 'message' => 'PHP ERROR: imagecopyresampled' ) );
-	          	exit;
-            }
+			if ($resampleReturn === false ) {
+				echo json_encode (array('status' => 'error', 'message' => 'PHP ERROR: imagecopyresampled' ) );
+				exit;
+			}
 
 			$imageSaveReturn = true;
-            if ($ext == "gif"){
-                $imageSaveReturn = imagegif($dst_img, $dst_file);
-            } else if($ext =="png"){
-                $imageSaveReturn = imagepng($dst_img, $dst_file);
-            } else {
-                $imageSaveReturn = imagejpeg($dst_img, $dst_file, $quality);
-            }
+			if ($ext == "gif"){
+				$imageSaveReturn = imagegif($dst_img, $dst_file);
+			} else if($ext =="png"){
+				$imageSaveReturn = imagepng($dst_img, $dst_file);
+			} else {
+				$imageSaveReturn = imagejpeg($dst_img, $dst_file, $quality);
+			}
 
-            if ($imageSaveReturn === false ) {
-            	echo json_encode (array('status' => 'error', 'message' => 'PHP ERROR: imagejpeg/imagegif/imagepng' ) );
-            	exit;
-        	}
-        }
+			if ($imageSaveReturn === false ) {
+				echo json_encode (array('status' => 'error', 'message' => 'PHP ERROR: imagejpeg/imagegif/imagepng' ) );
+				exit;
+			}
+		}
 
 		// Generate Retina Image
 		if( isset( $_POST['make2x'] ) && $_POST['make2x'] === 'true' ) {
@@ -296,19 +323,19 @@ setInterval(function() {
 						$img->set_quality( $quality );
 						$img->save($dst_file2x);
 					}else {
-            			echo json_encode (array('status' => 'error', 'message' => 'WP_ERROR: ' . $img->get_error_message() ) );
+						echo json_encode (array('status' => 'error', 'message' => 'WP_ERROR: ' . $img->get_error_message() ) );
 						exit;
-            		}
+					}
 				} else {
 					$dst_img2x = imagecreatetruecolor($dst_w2x, $dst_h2x);
 					$resampleReturn = imagecopyresampled($dst_img2x, $src_img, $dst_x, $dst_y, $src_x, $src_y, $dst_w2x, $dst_h2x, $src_w, $src_h);
 
-		            if ($resampleReturn === false ) {
-		            	echo json_encode (array('status' => 'error', 'message' => 'PHP ERROR: imagecopyresampled' ) );
-			          	exit;
-		            }
+					if ($resampleReturn === false ) {
+						echo json_encode (array('status' => 'error', 'message' => 'PHP ERROR: imagecopyresampled' ) );
+						exit;
+					}
 
-		            $imageSaveReturn = true;
+					$imageSaveReturn = true;
 					if ($ext == "gif"){
 						$imageSaveReturn = imagegif($dst_img2x, $dst_file2x);
 					} else if($ext =="png"){
@@ -316,7 +343,7 @@ setInterval(function() {
 					} else {
 						$imageSaveReturn = imagejpeg($dst_img2x, $dst_file2x, $quality);
 					}
-
+						
 					if ($imageSaveReturn === false ) {
 						echo json_encode (array('status' => 'error', 'message' => 'PHP ERROR: imagejpeg/imagegif/imagepng' ) );
 						exit;
